Changed byte to int in ContactType, Discount, and ContactStatus models.

+++ MOBYNew/Models/ContactType.cs
@@ -7,14 +7,14 @@ namespace MOBYNew.Models
 {
     public class ContactType
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
         public string contactType { get; set; }
 
         public Discount Discount { get; set; }
-        public byte DiscountId { get; set; }
+        public int DiscountId { get; set; }
         public bool? IsEligibleForDiscount { get; set; }
 
         public ContactStatus ContactStatus { get; set; }
-        public byte ContactStatusId { get; set; }
+        public int ContactStatusId { get; set; }
     }
 }
